Создает FilmsSectionController
